backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
confusable-homoglyphs==3.2.0
dataclasses==0.6
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-registration==3.0.1
django-rest-auth==0.9.5
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djongo==1.2.36
dnspython==1.16.0
gunicorn==19.9.0
idna==2.8
mysqlclient==1.4.4
oauthlib==3.1.0
Pillow==6.2.1
pipenv==2018.11.26
psycopg2==2.8.4
pycurl==7.43.0.3
PyJWT==1.7.1
pymongo==3.9.0
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.3
PyVimeo==1.0.11
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
tuspy==0.2.1
urllib3==1.25.6
virtualenv==16.7.5
virtualenv-clone==0.5.3
whitenoise==4.1.4


